refactor: extract cache logic to own file
